feat(prompt-node): editable prompt manager with version history
- Extract the chip editor into a reusable PromptEditor component (node passes variable/connected/create context; the manager uses it standalone).
- Make the manager dialog edit prompt content and create new prompts via the shared editor; saving an edit writes a new version, a name-only change updates metadata.
- Add a version history panel: restore an older version by saving it as the new latest.
